StandardCalc can evaluate simple addition expressions.
Added a getOperator method to OpStack that retrieves the plus symbol
operator dependent on the string variation given to it. Couldn't just
change Symbol enum as stopped the rest of the code working. Works by
stacking numbers and operators in two different stacks and then
iterating through these, popping into a string.